requirements.txt
Flask==3.0.3
gunicorn==22.0.0

Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.32
Flask-Migrate==4.0.7

Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2

python-dotenv==1.0.1

# PostgreSQL driver (Render-friendly)
pg8000==1.31.2

# Cloudflare R2 (S3)
boto3==1.35.10
botocore==1.35.10

pytz==2024.1
requests==2.34.0